Move Sigma<->Radius conversion to geometry
